requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
gpiozero
RPi.gpiozeropigpio
lgpio
python-crontab